server/requirements.txt
eth_account==0.5.6
fastapi==0.73.0
pydantic==1.9.0
PyJWT==2.3.0
pymongo==4.0.1
python-dotenv==0.19.2
python_bcrypt==0.3.2
requests==2.26.0
